+++ src/main.rs
@@ -1,7 +1,10 @@
 extern crate clap;
 extern crate fs;
+extern crate jsonschema;
 extern crate module;
 extern crate module_render;
+extern crate serde_json;
+extern crate thiserror;
 extern crate tracing;
 extern crate tracing_subscriber;
 
@@ -12,6 +15,7 @@ use clap::{Parser, Subcommand};
 
 use render::RenderArguments;
 use tracing_subscriber::{EnvFilter, FmtSubscriber};
+use validate::ValidateArguments;
 
 /// The entry point of the CLI execution.
 fn main() -> Result<(), Box<dyn Error>> {
@@ -27,6 +31,7 @@ fn main() -> Result<(), Box<dyn Error>> {
     match args.command {
         Commands::Build(args) => build::run(args),
         Commands::Render(args) => render::run(args),
+        Commands::Validate(args) => validate::run(args),
     }
 }
 
@@ -43,9 +48,12 @@ struct Cli {
 enum Commands {
     Build(BuildArguments),
     Render(RenderArguments),
+    Validate(ValidateArguments),
 }
 
 /// Implements the [Build](crate::Commands::Build) command.
 pub mod build;
 /// Implements the [Render](crate::Commands::Render) command.
 pub mod render;
+/// Implements the [Validate](crate::Commands::Validate) command.
+pub mod validate;